requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
PyYAML>=6.0

# Jetson GPIO
Jetson.GPIO>=2.0.17

# Serial communication
pyserial>=3.5

# Adafruit CircuitPython libraries
adafruit-circuitpython-dht>=3.7.0
adafruit-circuitpython-busdevice>=5.2.0
adafruit-blinka>=8.0.0

# Image processing and computer vision
Pillow>=9.0.0
scikit-image>=0.19.0

# AI/ML frameworks
torch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0

# TensorRT (Jetson optimized)
# Note: Install via apt on Jetson: sudo apt install python3-libnvinfer-dev

# ROS 2 dependencies (optional)
rclpy>=3.0.0; python_version>='3.8'
sensor-msgs>=4.0.0; python_version>='3.8'
geometry-msgs>=4.0.0; python_version>='3.8'

# HTTP and networking
requests>=2.28.0
websockets>=10.0

# Data handling
pandas>=1.5.0
matplotlib>=3.5.0

# Logging and configuration
loguru>=0.6.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
mock>=4.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# GPIO and hardware interfaces
RPi.GPIO>=0.7.1; platform_machine=="armv7l"
gpiozero>=1.6.0

# Additional sensor libraries
w1thermsensor>=2.0.0
board>=1.0
busio>=5.0.0

# Video and streaming
imageio>=2.22.0
imageio-ffmpeg>=0.4.0

# Utilities
click>=8.0.0
tqdm>=4.64.0
psutil>=5.9.0